snap: Rework into strict confinement
Make it possible to ship box64-with-gles as strictly confined.
It requires bundling a whole bunch of dependencies along with it, but also
has dependency on some commonly shared libraries from gnome extensions and
mesa-core22 content snaps, meaning less libraries to ship by itself.

Sadly though some applications might want to have a minimal execution environment
for scripts as well, so ship a few utilities with the stage-packages.